Move exchange logic to saga
